+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Cached configurations using `ImportIniConfig`.
 - Attribute access using the call directive (`call`).
+- An LRU cache `Stash` class.
 
 
 ## [1.14.3] - 2024-04-14